Major Update: Added role-based views and security
This commit implements the new SOAP calls in ng-sharepoint that allow
us to fetch the user info more reliably and also read what groups they
are members of.  From there, roles for the user are determined which
affect what the user can see or do.  Additional JS checks are in place
to prevent someone from arbitrarily editing the HTML/CSS.  Finally,
since we are only loading a snapshot of the server’s security groups,
the real security still exists server-side, we are just helping to
taylor the presentation for the user’s specific roles.