Removed Custom Filters Feature
This was an overly-complex system that while it gave the users more
searching power, it proved to be far too brittle with little
value-added.  Instead, much of the custom filters will now rely upon
smart archiving—showing the users only what they need to see and
treating everything else as if it is just archived data.