Redesign of requests view from wells to table
This was a major redesign but makes a lot more sense from an FTD
standpoint.  The well had a lot of superficial data that required extra
space and cluttered the interface.  This rework is all about
streamlining the FTD scheduler’s job of approving seat requests.  This
is also what drove the changes to the grouping/sorting options to make
the most of the interface.  We are going for optimal experience vs
overload of choices for this one (Apple vs Linux, basically).