Feature: Added [remember] attribute for selectize caching
This simple feature has a lot of potential to save time/frustration.
We can now intelligently auto-select the last used field, specified by
a shared name across the entire application.  For example
remember=‘ftd’ could be used on multiple pages and forms to save the
same data: the default FTD for this user.  Since this data is typically
the same across the entire app, it should save the users a lot of time
and is transparent in operation since it always remembers the last-used
value anyway.  This is one of those things I’ve been wanting to do for
quite some time.