Split scheduled view into search and manage views
Really tried to keep everything on one view for simplicities sake but
in the end, multiple views well provide a better overall user
experience and the issues of not being DRY and sharing unauthorized
pages are easy kills.  A new Jade mixin was created for these two views
since they are so similar.  When the app is complied it will mean
duplicate HTML/CSS, but on the development side it’s just one codebase.
 Also, a simple REGEX can be used to force sharing the /scheduled-ftd/
page as /scheduled/ since the tags and searches are preserved anyway it
should be transparent to the users.