Finished modal calendar course add/update
The class start/stop fields were completely removed and now a scheduler
just clicks the day they want the class to start.  FTSS will pre-fill a
range based on the class day length setting (skips weekends
automatically).

Also allowed the updating of existing classes view drag or slide.  Had
to make some changes to how the dates are saved, in FullCalendar, the
last day of an all-day event automatically is set for 00:00 the next
day.  So a class ending on a Friday would display as Saturday at 00:00.
 This is pretty annoying but since it is at least consistent on the
calendar views, we’ll just update our rows during preprocessing to
account for it.

This should resolve the timezone issues and date offsets occurring
previously.