Major feature - built gantt-style scheduler view
FullCalendar is great for giving us a calendar view, but classes became
had to read as the schedule filled up, this new horizontal view
provides much better readability and is similar to how most FTDs view
class schedules now.  This will likely become the primary interface for
FTD schedulers to manage their classes.

Eventually this will be rolled into a “scheduler” directive so that
instructors or others can see a read-only view of this data as well.