Added secondary row to gantt chart to give space/details
This is just a visual spacer that will also serve as the buffer between
rows to show details on hover.  Also removed the first column that had
course info as it was of little practical value without the connected
data (start/stop times) and I didn’t want to have to make two tables
out of this.