****Feature Deprecation
Removed user-driven group-by and sort-by options.  These really weren’t
that useful and many pages had no options leading to a confusing and
crowded interface.  Now each page is optimized for the best
grouping/sorting for that particular view and the user can leverage the
instant-search feature to find exactly what they are looking for.  This
was also driven by a desire to simplify the core code and make it more
reliable.