Working drag & drop directly onto bio photo for inline updates
To complete the inline updates for instructor data, this is a final
push to get drag & drop photos to work for bio photos inline.  It
mostly works now but there are a few outstanding issues:

1.  Need to add CSS to resize the drag & drop user cue for the smaller
drop target
2.  Bio photos are not refreshing properly after the update completes
(need to check the photo directive’s watch or ng-repeat’s track by
3.  Our submit spinner doesn’t seem to be updating as expected….don’t
really know why on that one.