Fixed offline mode user group issue
This was an intermittent issue anyway with user/group loading.  Having
the watch loaded after the $scope load function (SharePoint.user())
posed the risk of cached responses causing the page to fail to load.
Basically, the $scope.$watch was being bound AFTER the variable called
and when the response was instant, it was causing the page to not
finish loading.  *fail*