Finished bio photo inline drag & drop update
Had to go with a jQuery workaround in the controller (yes, bad form)
due to the way that ng-repeat watches collections.  Basically, updates
were not being passed back to the view with ng-repeat so we used jQuery
to toggle the spinner visibility and update the bio photo once the new
one is uploaded to SP.

Also removed the hard-coded URLs and used the config urls instead (one
less place to have to change URLs).