Added "% Annual Instructor Utilization" to stats view
This is a an interesting piece of data that we are going to highlight
on the instructor stats view.  Basically, we count how many hours they
taught over the last year in FTSS and divide that by 48 work weeks
(19.2 / 1920) to give a % of the average work year spent actually
teaching.  We take three weeks off of each year for this calculation to
account for Christmas break, federal holidays and AETC down days.