*smarter* Take advantage of cached schedule data
Just like the rewritten manage-ftd page, this page no longer performs
the needless calls back to sharepoint for instructor schedules.  Since
splitting the schedule page to an FTD-only page, we’ve always had the
cached data anyway, now we are taking more advantage of that data to
simplify the program and speed things up even more.