Fixed empty strings being passed as data
This was breaking some updates, i.e. no instructor on build class.
Basically, ng-SharePoint was detecting a value (empty String) and
sending that up, but the field was actually a number field so it was
causing the create/update requests to fail.  Null values are ignored by
ng-SharePoint to avoid the issue.