Fixed "undefined" instructor bug
Really need to look at this one again later on.  This is more of a
workaround than anything.  It’s not a bad one, but the issue of
caches.Instructors missing labels and other data built during Selectize
initialization is concerning.  The issue occurs when navigating from
the managed-ftd screen to the scheduled-ftd screen.