Closed #4, add Training Session capability
This is actually just a very simple hack, but really meets the intent
without added complexity.  Also added some user shortcuts for leave,
old classes and training sessions.  Biggest issue will be when class #s
are added by an FTD later on—if they delete the TS from the field, it
will no longer display as a training session.