Created row.dateRange, removed row.startText/row.endText
This was added several extra calls across the app so was moved up to
the controller instead.  This will also give us the flexibility to
tailor the range a little more.  The first change was for single-day
classes, now they just show the day of the class rather than duplicate
days.  Example:

5 Jan 2015 - 5 Jan 2015 becomes just 5 Jan 2015 everywhere.