Refactored FullCalendar controller code
Optimized this messy code.  Also, fixed the issue where courses would
disappear when you add a class to the schedule (only affected the modal
view when building classes).

The other important usability change is the removal of the duplicate
course from the instructor’s schedule view when building classes.  This
was really confusing for uses because it statically displayed a
duplicate class for the current class.