Resource view refactor into directive
This is a major change to the resource view.  The scope was cleaned up
a lot, removed nearly all Angular work and pushed it to plain JS string
concatenation for speed.  Also did some minor layout/style enhancements
including adding a footer to the view for large lists.  I first added a
modulas to show the days/months every X number of lines, but it looked
cluttered.