Breaking schema change (again)
Thank you SP for being lame.  Now we will use a simple integer offset
to track the start of classes.  It will be the # of days since Y2K
(seemed fitting).  Since we want times to “magically” treat all times
as local (all classes are transcribed in Central timezone for their
respective local timezones).  This is a hugely confusing issue for the
software so just using a day-offset lets it always seem to be local to
the current user.