Added Course/Unit/Instructor prefetch back
SP was having performance issues with the initial load of the
course/instructor lists.  I did some experiments with not loading the
entire instructor list, but thanks to SP *again* failing in some weird
way that’s not working.  Basically, because we make the instructors
optional vs mandatory, SP kills over when using the $expand oData
feature….so for now we’ll just do some prefectch the help out SP.
Added bonus though in that prefetch not only pre-parses, but also lets
us pre-GZIP the data one time so this should be a major performance
boost for the first execution.  This also means that a stale cache is
less of an issue as it is so much less data anyway.