MTT Feature DONE--finally! Also added class build chooser
This one took a lot of work, but now an FTD can ID a class as MTT and
it will show up as it should pretty much everywhere.  The shortcuts
were also replaced with a dropdown to choose MTT vs TS vs Leave, etc.
This really makes more sense and should alleviate some confusion in
building classes.