refactor: close #25, redesigned data schema for TS/NA/MTT tracking
This is a breaking change that should stabilize several areas of FTSS.
All the hacks against CourseId and TTMS fields have now been removed
and we are using .MTT, .NA and .TS to identify the special use-cassis.
Also, “Old” was dropped as J4 won’t see old classes anyway now.

- Added MTT location to resource view
- Changed seat # from open to student count
- Cleaned up modal-scheduled UI some
- Implemented more ng-if’s to speed things up
- Disabled changing course #/type after class is created
- Added missing user instructions
- Cut a lot of fat out of the modal controller code
- Removed fake course from selectize as this is no longer needed (and
was buggy anyway)
- Increased width of course selector to prevent word wrap