refactor (major): moving requests back to their own list for editing/canceling/requirements
This is major schema change (back to the original db design on this
one) for the student requests.  Due to the increasing complexity/need
for editing of students and integration of GAL lookups, Requests_JSON
was pulled out into the new Requests list.  Major points below:

- Each request is a single item
- Only JSON is Students_JSON which is an array of name:email items
- Approved is an integer stored with the class now (we will be use
SharePoint.batch() a lot now)
- Viewing students will requiring a secondary async call (important
because this means we don’t have to load all that JSON data anymore for
looking at schedules)
- Host and Unit are marked in the request for optimized lookups (we
don’t bind to the class w/$extend due to the needlessly complexity it
would require)
- This will enable much safer/simpler seat cancellation/edits
- Could eventually be isolated to enable PII storage as it is a
separate list/can have service side controls
- Support for GAL lookups now functional
- Greatly simplified overall code for requests, things just make more
sense now
- Did have to remove pending seats tracking in other places (this isn’t
really that big of a deal as the place to look for this is the Requests
pages anyway)
- Sets us up for a much simpler solution to the 898/requirements piece
(just add requests without a ClassId assigned and then assign the
ClassId when building the class)