refactor: breaking change - migrated photos to CDN vs SP
This is a major change (and not complete), due to SP being super crappy
about caching, we’re moving all the bio photos to AWS.  This shouldn’t
be an issue as no PII is associated with the images, they are based off
of random GUIDs and we will restrict access to the images though AWS
access controls.  The major benefit is native browser caching of
images, combined with a smarter uploader (will be implemented before
publishing), should fix the long standing issue of overside files being
uploaded.