refactor: major page init/routing rewrite
Trying to reduce the complexity of the page load functions and wasted
processing/parsing that was occurring when we still passed around a
share url.  Much of the code was needed to support URL sharing across
views, now this is done through other mechanisms (ftp-centric
view/host-centric view/selectize remember attributes).  This is a work
in prorgress, but should result in cleaner URLs, faster page loads and
more reliable data parsing.