refactor: Moved selectize into app
This was becoming a continual problem.  We’re using a highly-customized
version of SelectizeJS that is several version old now.  Most of the
customization has to do with performance and only a few of these
changes have been pushed to the parent repo.  Since it is so customized
and still highly performant, we will just keep it local instead and cut
out any of the extra stuff not needed as we move forward.