refactor: more work on the production view/moved back to caching model
Had to fix a couple issues with watchers (removed some bind-once
params) and normalized the caching for optimal rending.  Also
implemented standard controller-helper service the same way that
scheduled-ftp binds for simplicity.  The graphs and inline graphs (per
instructor) should now be in sync and not offset by a month or so.
Finally, removed current months data from this view as it would throw
off the graphs for the current month.