feature: Added FTD scheduler view for MTS with a dedicated FTD
This is a read-only view that is customized for host units to see all
the FTD training activity quickly and easily.  They will get the same
benefits that FTDs get with the exception of some instructor-specific
data removed (such as leave/TDY details).  This will also let them do a
one-click look at students and make requests for out of cycle training
needs.