feature: added independent unavailability tracking
This is to fix a major SharePoint indexing issue with yes/no fields and
to reduce the size of the Scheduled list.  Instructors will now be able
to create their own leave/TDY or other unavailability in FTSS while
still keeping the original interface for schedulers to create items for
anyone.  The changes across the class model are significant and a new
combining function was written in classProcessor to handle this new
challenge.