Feat: moved FTSS update messages to SP with ftss-updates directive to keep the view clean
This is the first of several moves to SP to thin the code, but also make
FTSS more dynamic/maintainable for future changes.  These changes are
now appropriate since we have the transparent batch operations to bundle
the requests into a single HTTP call.  One of the ways we'll keep the
load low is caching, but also moving instructors into a JSON object of
the unit will offset the server load.