Refactor: Breaking change--moved geodata, language and downDays data to new AppAssets SP list
So this will be an important change for the long-term viability of FTSS.
Once again taking advantage of SP batch processing, we have a new list
call AppAssets that stores JSON objects for FTSS that may change over
time and shouldn't require a code change to show up.

This will allow non-developer FTSS team members to make changes to FTSS
in important ways that could only be done through code previously.
There may still be some cleaning up to do on the language file, so more
commits are likely for that.