refactor: breaking change- removed private options variable, merged into caches
This was a carry-over from when the main class search selectize object
was used on every page for nearly every filter--now that it is only used
on the class search page, it made sense to merge these into one object
again.  It was also preventing us from using custom Host lists like the
other lists.